update products and reviews routes/controllers

+++ back-end/src/routes/products/products.js
@@ -4,6 +4,7 @@ app.use(express.json())
 
 
 const Products = require('../../models/products/products')
+const User = require('../../models/users/user')
 
 const GetProducts = async (req,res) => {
     const product  = await Products.find()
@@ -18,9 +19,19 @@ const GetProductsById = async (req,res) => {
 
 const PostProducts = async (req,res)=>{
     try{
+        
+        const admin = await User.findById(req.user.user_id)
+
+        if(!admin.roles.includes('Admin')){
+            return res.status(400).json({
+                message:"No estas autorizado para hacer esto :/"
+            })
+        }
+
         const product = new Products(req.body)
         await product.save()
         res.json(product)
+
     }catch(err){
         res.json(err)
     }
@@ -28,10 +39,20 @@ const PostProducts = async (req,res)=>{
 
 const UpdateProduct = async (req,res)=>{
     try{
+
+        const admin = await User.findById(req.user.user_id)
+
+        if(!admin.roles.includes('Admin')){
+            return res.status(400).json({
+                message:"No estas autorizado para hacer esto :/"
+            })
+        }
+
         const {id} = req.params
         const UpdProduct = req.body
         await Products.findByIdAndUpdate(id, UpdProduct)
         res.json(UpdProduct)
+
     }catch(err){
         res.json(err)
     }
@@ -39,9 +60,19 @@ const UpdateProduct = async (req,res)=>{
 
 const DeleteProducts = async (req,res) => {
     try{
+
+        const admin = await User.findById(req.user.user_id)
+
+        if(!admin.roles.includes('Admin')){
+            return res.status(400).json({
+                message:"No estas autorizado para hacer esto :/"
+            })
+        }
+
         const {id} = req.params
         const product  = await Products.findByIdAndRemove(id)
         res.json(product)
+
     }catch(err){
         res.json(err)
     }
